// src/app/modules/admin/model/podcasts.ts
export class PodcastsData {
    PodcastID: any;
    Title: any;
    PodcastType: boolean;
    Description: any;
    Embed_Code: any;
    Image: any;
    Published: any;
    ResourceUrl: any;
    Slug: any;
    Source: any;
    SortOrder: any;
    IsActive: any;
    StartDate: any;
    StartTime: any;
    EndDate: any;
    EndTime: any;
    CreatedBy: any;
    CreatedDate: any;
    ModifiedBy: any;
    ModifiedDate: any;
    podcastInfo: any;
    PublishedDate: any;

}